# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateMemoryRequest": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "reason": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "content",
          "reason"
        ],
        "title": "CreateMemoryRequest",
        "type": "object"
      },
...
      "DeleteMemoryRequest": {
        "additionalProperties": false,
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          }
        },
        "required": [
          "expected_version"
        ],
        "title": "DeleteMemoryRequest",
        "type": "object"
      },
      "ErrorDetailResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDetailResponse",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetailResponse"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "MarketDetailsResponse": {
        "additionalProperties": false,
        "properties": {
          "breadth": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Breadth"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/MarketResourceErrorResponse"
            },
            "title": "Errors",
            "type": "array"
          },
          "flash_news": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Flash News",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "headlines": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Headlines",
            "type": "array"
          },
          "hotspots": {
            "additionalProperties": {
              "items": {
                "additionalProperties": true,
                "type": "object"
              },
              "type": "array"
            },
            "title": "Hotspots",
            "type": "object"
          },
          "rankings": {
            "additionalProperties": {
              "items": {
                "additionalProperties": true,
                "type": "object"
              },
              "type": "array"
            },
            "title": "Rankings",
            "type": "object"
          },
          "turnover": {
            "additionalProperties": true,
            "title": "Turnover",
            "type": "object"
          }
        },
        "required": [
          "generated_at",
          "turnover",
          "rankings",
          "hotspots",
          "headlines",
          "flash_news",
          "errors"
        ],
        "title": "MarketDetailsResponse",
        "type": "object"
      },
      "MarketIndicesResponse": {
        "additionalProperties": false,
        "properties": {
          "errors": {
            "items": {
              "$ref": "#/components/schemas/MarketResourceErrorResponse"
            },
            "title": "Errors",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "indices": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Indices",
            "type": "array"
          },
          "trends": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Trends",
            "type": "array"
          }
        },
        "required": [
          "generated_at",
          "indices",
          "trends",
          "errors"
        ],
        "title": "MarketIndicesResponse",
        "type": "object"
      },
      "MarketOverviewResponse": {
        "additionalProperties": false,
        "properties": {
          "breadth": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Breadth"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/MarketResourceErrorResponse"
            },
            "title": "Errors",
            "type": "array"
          },
          "flash_news": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Flash News",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "headlines": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Headlines",
            "type": "array"
          },
          "hotspots": {
            "additionalProperties": {
              "items": {
                "additionalProperties": true,
                "type": "object"
              },
              "type": "array"
            },
            "title": "Hotspots",
            "type": "object"
          },
          "indices": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Indices",
            "type": "array"
          },
          "rankings": {
            "additionalProperties": {
              "items": {
                "additionalProperties": true,
                "type": "object"
              },
              "type": "array"
            },
            "title": "Rankings",
            "type": "object"
          },
          "trends": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Trends",
            "type": "array"
          },
          "turnover": {
            "additionalProperties": true,
            "title": "Turnover",
            "type": "object"
          }
        },
        "required": [
          "generated_at",
          "indices",
          "trends",
          "turnover",
          "rankings",
          "hotspots",
          "headlines",
          "flash_news",
          "errors"
        ],
        "title": "MarketOverviewResponse",
        "type": "object"
      },
      "MarketResourceErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "item_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Id"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "resource": {
            "title": "Resource",
            "type": "string"
          }
        },
        "required": [
          "resource",
          "item_id",
          "message"
        ],
        "title": "MarketResourceErrorResponse",
        "type": "object"
      },
      "MemoryActivityResponse": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "memory_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Id"
          },
          "operation": {
            "enum": [
              "read",
              "create",
              "update",
              "delete"
            ],
            "title": "Operation",
            "type": "string"
          },
          "result_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Count"
          },
          "task_id": {
            "title": "Task Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "operation",
          "memory_id",
          "content",
          "result_count",
          "task_id",
          "created_at"
        ],
        "title": "MemoryActivityResponse",
        "type": "object"
      },
...
      "MemoryItemResponse": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_task_id": {
            "title": "Created Task Id",
            "type": "integer"
          },
          "deleted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "updated_task_id": {
            "title": "Updated Task Id",
            "type": "integer"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "content",
          "reason",
          "created_task_id",
          "updated_task_id",
          "version",
          "created_at",
          "updated_at",
          "deleted_at"
        ],
        "title": "MemoryItemResponse",
        "type": "object"
      },
      "MemoryOverviewResponse": {
        "additionalProperties": false,
        "properties": {
          "activities": {
            "items": {
              "$ref": "#/components/schemas/MemoryActivityResponse"
            },
            "title": "Activities",
            "type": "array"
          },
          "activity_total": {
            "title": "Activity Total",
            "type": "integer"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "item_match_total": {
            "title": "Item Match Total",
            "type": "integer"
          },
          "item_total": {
            "title": "Item Total",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/MemoryItemResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "activities",
          "activity_total",
          "items",
          "item_total",
          "item_match_total",
          "generated_at"
        ],
        "title": "MemoryOverviewResponse",
        "type": "object"
      },
...
      "UpdateMemoryRequest": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "reason": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "expected_version",
          "content",
          "reason"
        ],
        "title": "UpdateMemoryRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AniuSessionCookie": {
        "in": "cookie",
        "name": "aniu_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "基于股票交易的可进化智能体系统",
    "title": "Aniu",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/aniu/market/overview": {
      "get": {
        "operationId": "get_market_overview_api_aniu_market_overview_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketOverviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request forbidden"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request rate limited"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Upstream integration failed"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "AniuSessionCookie": []
          }
        ],
        "summary": "Get Market Overview",
        "tags": [
          "Market"
        ]
      }
    },
    "/api/aniu/market/overview/details": {
      "get": {
        "operationId": "get_market_details_api_aniu_market_overview_details_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketDetailsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request forbidden"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request rate limited"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Upstream integration failed"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "AniuSessionCookie": []
          }
        ],
        "summary": "Get Market Details",
        "tags": [
          "Market"
        ]
      }
    },
    "/api/aniu/market/overview/indices": {
      "get": {
        "operationId": "get_market_indices_api_aniu_market_overview_indices_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketIndicesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request forbidden"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request rate limited"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Upstream integration failed"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service unavailable"
          }
        },
        "security": [
          {
            "AniuSessionCookie": []
          }
        ],
        "summary": "Get Market Indices",
        "tags": [
          "Market"
        ]
      }
    },
    "/api/aniu/memories": {
      "get": {
        "operationId": "get_memory_overview_api_aniu_memories_get",
        "parameters": [
          {
            "in": "query",
            "name": "activity_limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Activity Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "activity_offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Activity Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "activity_task_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Activity Task Id"
            }
          },
          {
            "in": "query",
            "name": "activity_operation",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "read",
                    "create",
                    "update",
                    "delete"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Activity Operation"
            }
          },
          {
            "in": "query",
            "name": "item_limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Item Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "item_offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Item Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "item_keywords",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 200,
              "title": "Item Keywords",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryOverviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "AniuSessionCookie": []
          }
        ],
        "summary": "Get Memory Overview",
        "tags": [
          "Memories"
        ]
      },
      "post": {
        "operationId": "create_memory_api_aniu_memories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "AniuSessionCookie": []
          }
        ],
        "summary": "Create Memory",
        "tags": [
          "Memories"
        ]
      }
    },
    "/api/aniu/memories/{memory_id}": {
      "delete": {
        "operationId": "delete_memory_api_aniu_memories__memory_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "title": "Memory Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteMemoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "AniuSessionCookie": []
          }
        ],
        "summary": "Delete Memory",
        "tags": [
          "Memories"
        ]
      },
      "put": {
        "operationId": "update_memory_api_aniu_memories__memory_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "title": "Memory Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "security": [
          {
            "AniuSessionCookie": []
          }
        ],
        "summary": "Update Memory",
        "tags": [
          "Memories"
        ]
      }
    },
...
  }
}